refactor(entity): DB에서 불러오는 객체를 Entity로 분리
