+++ src/main/java/entity/GameEntity.java
@@ -1,17 +1,17 @@
-package dto.dao;
+package entity;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 
-public record LoadedGameState(long gameId, double choScore, double hanScore, String turnTeam,
-                              List<LoadedPiece> pieces) {
-    public static LoadedGameState fromHeaderResultSet(long gameId, ResultSet headerRs, List<LoadedPiece> pieces)
+public record GameEntity(long gameId, double choScore, double hanScore, String turnTeam,
+                         List<PieceEntity> pieces) {
+    public static GameEntity fromHeaderResultSet(long gameId, ResultSet headerRs, List<PieceEntity> pieces)
             throws SQLException {
         if (!headerRs.next()) {
             throw new IllegalStateException("재개할 수 없는 게임입니다: id=" + gameId);
         }
-        return new LoadedGameState(
+        return new GameEntity(
                 gameId,
                 headerRs.getBigDecimal("cho_score").doubleValue(),
                 headerRs.getBigDecimal("han_score").doubleValue(),
